+++ src/config/constants.ts
@@ -5,10 +5,12 @@ import importedSett from '../config/system/abis/Sett.json';
 import importedGeyser from '../config/system/abis/BadgerGeyser.json';
 import importedBscErc20 from '../config/system/abis/BscErc20.json';
 import importedYearnSett from '../config/system/abis/YearnWrapper.json';
+import importedGuestList from '../config/system/abis/GuestList.json';
 import { tokens, sett_system } from './deployments/mainnet.json';
 import { NetworkConstants, ClaimsSymbols } from '../mobx/model';
 import { getNetworkDeploy } from 'mobx/utils/network';
 import { AbiItem } from 'web3-utils';
+import { PartialAttemptOptions } from '@lifeomic/attempt';
 
 export const RPC_WALLETS: { [index: string]: boolean } = {
 	ledger: true,
@@ -62,6 +64,8 @@ export const CLAIMS_SYMBOLS: ClaimsSymbols = {
 		[sett_system.vaults['native.badger']]: 'bBadger',
 		[sett_system.vaults['native.digg']]: 'bDigg',
 		[tokens.defiDollar]: 'Defi Dollar',
+		[sett_system.vaults['native.cvx']]: 'bCVX',
+		[sett_system.vaults['native.cvxCrv']]: 'bcvxCRV',
 	},
 };
 
@@ -71,23 +75,26 @@ export const ERC20_ABI = importedErc20.abi as AbiItem[];
 export const SETT_ABI = importedSett.abi as AbiItem[];
 export const YEARN_ABI = importedYearnSett.abi as AbiItem[];
 export const GEYSER_ABI = importedGeyser.abi as AbiItem[];
+export const GUEST_LIST_ABI = importedGuestList.abi as AbiItem[];
 export const BSC_ERC20 = importedBscErc20;
 export const APP_NAME = 'badgerDAO';
 export const PORTIS_APP_ID = 'cbf7534d-170d-4903-943f-e607dc588b7f';
 export const EMPTY_DATA = '0x';
 export const ZERO_CURRENCY = '0.00000';
-export const SITE_VERSION = 'v2.7.6';
+export const SITE_VERSION = 'v2.8.0';
 export const WC_BRIDGE = 'https://wc-bridge.badger.finance/';
 
-const toBool = (val: string | undefined): boolean => (val ? val === 'true' : false);
+const toBool = (val: string | undefined): boolean => (val ? val.toLowerCase() === 'true' : false);
 
 export const FLAGS = {
-	IBBTC_FLAG: toBool(process.env.REACT_APP_IBBTC_FLAG),
 	WBTC_FLAG: toBool(process.env.REACT_APP_BRIDGE_WBTC),
 	GEYSER_FLAG: toBool(process.env.REACT_APP_GEYSER_ENABLED),
 	EXPERIMENTAL_VAULTS: toBool(process.env.REACT_APP_EXPERIMENTAL_VAULTS),
 	BOOST_V2: toBool(process.env.REACT_APP_BOOST_V2),
 	IBBTC_OPTIONS_FLAG: toBool(process.env.REACT_APP_IBBTC_OPTIONS_FLAG),
+	CONVEX_SETTS: toBool(process.env.REACT_APP_CONVEX_SETTS),
+	STABILIZATION_SETTS: toBool(process.env.REACT_APP_STABILIZATION_SETTS),
+	BOOST_OPTIMIZER: toBool(process.env.REACT_APP_BOOST_OPTIMIZER),
 };
 
 export const ZERO = new BigNumber(0);
@@ -98,6 +105,18 @@ export const CURVE_WBTC_RENBTC_TRADING_PAIR_ADDRESS = '0x93054188d876f558f4a66B2
 export const RENVM_GATEWAY_ADDRESS = '0xe4b679400F0f267212D5D812B95f58C83243EE71';
 export const RENVM_NETWORK = 'mainnet';
 export const DEBUG = process.env.NODE_ENV !== 'production';
+export const ZERO_ADDR = '0x0000000000000000000000000000000000000000';
 
 // time constants
 export const ONE_MIN_MS = 60 * 1000;
+export const baseRetryOptions = {
+	// delay defaults to 200 ms.
+	// delay grows exponentially by factor each attempt.
+	factor: 1.5,
+	// delay grows up until max delay.
+	maxDelay: 1000,
+	// maxAttempts to make before giving up.
+	maxAttempts: 3,
+};
+export const defaultRetryOptions: PartialAttemptOptions<void> = baseRetryOptions;
+export const getDefaultRetryOptions = <T>(): PartialAttemptOptions<T> => baseRetryOptions;