+++ src/mobx/model/vault-cap copy.ts
@@ -1,0 +1,5 @@
+import { VaultCap } from './vault-cap';
+
+export interface VaultCaps {
+	[address: string]: VaultCap;
+}