+++ src/mobx/model/vault-cap.ts
@@ -1,0 +1,9 @@
+import { TokenBalance } from './token-balance';
+
+export interface VaultCap {
+	vaultCap: TokenBalance;
+	totalVaultCap: TokenBalance;
+	userCap: TokenBalance;
+	totalUserCap: TokenBalance;
+	asset: string;
+}