+++ src/mobx/utils/apiV2.ts
@@ -7,7 +7,9 @@ import {
 	Account,
 	RewardMerkleClaim,
 	LeaderBoardData,
+	LeaderBoardEntry,
 } from 'mobx/model';
+import { TokenConfig } from 'mobx/model/token-config';
 
 export const getApi = (): string => {
 	if (process.env.REACT_APP_BUILD_ENV === 'production') {
@@ -19,6 +21,7 @@ const badgerApi = getApi();
 
 // api endpoints
 const listSettsEndpoint = `${badgerApi}/setts`;
+const getTokensEndpoint = `${badgerApi}/tokens`;
 const getPricesEndpoint = `${badgerApi}/prices`;
 const getTVLEndpoint = `${badgerApi}/value`;
 const checkShopEndpoint = `${badgerApi}/reward/shop`;
@@ -30,6 +33,10 @@ const getLeaderBoardDataEndpoint = `${badgerApi}/leaderboards`;
 // api function calls
 export const listSetts = async (chain?: string): Promise<Sett[] | null> => {
 	return fetchData(() => fetch(`${listSettsEndpoint}${chain ? `?chain=${chain}` : ''}`));
+};
+
+export const getTokens = async (chain?: string): Promise<TokenConfig | null> => {
+	return fetchData(() => fetch(`${getTokensEndpoint}${chain ? `?chain=${chain}` : ''}`));
 };
 
 export const getTokenPrices = async (chain?: string, currency?: string): Promise<PriceSummary | null> => {
@@ -60,6 +67,10 @@ export const fetchClaimProof = async (address: string): Promise<RewardMerkleClai
 
 export const fetchLeaderBoardData = async (page: number, size: number): Promise<LeaderBoardData | null> => {
 	return fetchData(() => fetch(`${getLeaderBoardDataEndpoint}?page=${page}&size=${size}`));
+};
+
+export const fetchCompleteLeaderBoardData = async (): Promise<LeaderBoardEntry[] | null> => {
+	return fetchData(() => fetch(`${getLeaderBoardDataEndpoint}/complete`));
 };
 
 const fetchData = async <T>(request: () => Promise<Response>): Promise<T | null> => {